refactor: split session party and encounter features
